Room dependency injection with dagger and hilt
Creating a task dao to insert, get and delete tasks. A room database with a
callback that inserts sample tasks into the database. An app module that is a
container for the database which makes it easier to inject and use the taskdao
throughout the app. Root application file to make use of dagger easier.
Finally, injecting taskdao into the TasksViewModel and setting up TasksFragment
to be used with dagger.