Swipe to delete and undo functionality
Adding swipe to delete on UI and view model now implements a task event channel
which acts as an queue to hold undo tasks event. This is exposed to the
fragment as a flow and is recieved by the fragment toshow the snackbar and undo
the deletion implemented in the view model.